complete sidebar functionality
